chore: remove non layout code from default layout
Cleanup default.vue layout, it included lost of global setup/state management code.
If we want multiple layouts this global setup/state needs to be moved outside of the layout
